requirements.txt
django
djangorestframework
djangorestframework-gis
psycopg2-binary
django-filter
